Use Pimcore\Tool::classExists instead of class_exists in OrderManager
While trying to find a suitable class name for order and orderItem
classes, class_exists() was used to find an existing class name, however
the Zend autoloader triggers a warning when a class is not found
resulting in alternative names (e.g. \\Listing -> _List) never being
tested as the execution stops before when error handling is set to E_ALL.

Pimcore\Tool::classExists turns off those errors while testing for
existing classes which is what we want in this case.