Update config.yaml
When using Symfony Encore, asset filenames can have versioning file suffixes. When using the symfony/asset twig extension, the files can be found without versioning suffix (usually in DEV mode). When versioning is enabled (PROD environment) the files cannot be located anymore, when the location of the manifest.json is not configured.